Make API more flexible
This changes will make API more flexible by allowing each API to use
`async with` style.
