+++ src/paspybin/api/api.py
@@ -1,3 +1,7 @@
+from traceback import TracebackException
+from types import TracebackType
+from typing import Self
+
 from aiohttp import ClientSession
 
 from ..types import DevKey, UserKey
@@ -30,6 +34,21 @@ class API:
         self._dev_key = dev_key
         self._user_key = user_key
         self._session = session if session is not None else ClientSession(self.base_url)
+
+    async def __aenter__(self) -> Self:
+        return self
+
+    async def __aexit__(
+        self,
+        exc_type: Exception,
+        exc_val: TracebackException,
+        traceback: TracebackType,
+    ) -> None:
+        await self.close()
+
+    async def close(self) -> None:
+        if self._session is not None:
+            await self._session.close()
 
     def is_authenticated(self) -> bool:
         return self._user_key is not None
